# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Akkudoktor API",
    "description": "Willkommen bei der API von Akkudoktor. Diese API stellt Ihre Daten ist nach der OpenAPI 3.0 Spezifikation bereit und ist aktuell kostenlos. \n\nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.\n\n_If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_\n\nSome useful links:\n- [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)\n- [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@akkudoktor.net"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "externalDocs": {
    "description": "Sourcecode",
    "url": "https://github.com/nick81nrw/solApi"
  },
  "servers": [
    {
      "url": "https://api.akkudoktor.net/"
    }
  ],
  "tags": [
    {
      "name": "forecast",
      "description": "pv generation forcast",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "paths": {
    "/forecast": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Get forcast data",
        "description": "get pv power from a location ",
        "operationId": "getForecast",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "latitude for location",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 55.45
          },
          {
            "name": "lon",
            "in": "query",
            "description": "longitude for location",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 8.14
          },
          {
            "name": "power",
            "in": "query",
            "description": "power pv modules in watts",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 3400
          },
          {
            "name": "azimuth",
            "in": "query",
            "description": "orientation of pv modules in azimuth in degree\n0 = south\n-45 = south-east\n-90 = east\n45 = south-west\n90 = west",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": -30
          },
          {
            "name": "tilt",
            "in": "query",
            "description": "tilt of pv modules degree\n0 = flat / horizontal\n30 = 30° tilt (on a roof)\n90 = vertical (at a wall)",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 35
          },
          {
            "name": "cellCoEff",
            "in": "query",
            "description": "cellcoefficient of your pv modules. It is used for calculate losses with the temperature of the modules. Value in percent. default: -0.4",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": -0.36
          },
          {
            "name": "timezone",
            "in": "query",
            "description": "timezone. default: Europe/Berlin",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Europe/Berlin"
          },
          {
            "name": "powerInvertor",
            "in": "query",
            "description": "max power of the invertor in watts. it will used, if the power of the invertor is lower than the power generation of the modules. default: no limit",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 4500
          },
          {
            "name": "inverterEfficiency",
            "in": "query",
            "description": "inverter efficiency. default: 1 (=100%)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 0.95
          }
        ]
      }
    },
    "/archive": {
      "get": {
        "tags": [
          "archive"
        ],
        "summary": "Get archive data",
        "description": "get pv power from a location ",
        "operationId": "getArchive",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "latitude for location",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 55.45
          },
          {
            "name": "lon",
            "in": "query",
            "description": "longitude for location",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 8.14
          },
          {
            "name": "power",
            "in": "query",
            "description": "power pv modules in watts",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 3400
          },
          {
            "name": "azimuth",
            "in": "query",
            "description": "orientation of pv modules in azimuth in degree\n0 = south\n-45 = south-east\n-90 = east\n45 = south-west\n90 = west",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": -30
          },
          {
            "name": "tilt",
            "in": "query",
            "description": "tilt of pv modules degree\n0 = flat / horizontal\n30 = 30° tilt (on a roof)\n90 = vertical (at a wall)",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 35
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "date from which the data sould load. default 7 days before today",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "2023-05-06"
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "date until which the data sould load. default yesterday",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "2023-05-10"
          },
          {
            "name": "cellCoEff",
            "in": "query",
            "description": "cellcoefficient of your pv modules. It is used for calculate losses with the temperature of the modules. Value in percent. default: -0.4",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": -0.36
          },
          {
            "name": "timezone",
            "in": "query",
            "description": "timezone. default: Europe/Berlin",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Europe/Berlin"
          },
          {
            "name": "powerInvertor",
            "in": "query",
            "description": "max power of the invertor in watts. it will used, if the power of the invertor is lower than the power generation of the modules. default: no limit",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 4500
          },
          {
            "name": "inverterEfficiency",
            "in": "query",
            "description": "inverter efficiency. default: 1 (=100%)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "example": 0.95
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
    "/prices": {
      "get": {
        "tags": [
          "prices"
        ],
        "summary": "Get energy price data",
        "description": "get price data for a defined range",
        "operationId": "getPrices",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "description": "start date or datetime. default: default: today 0 am",
            "required": false,
            "schema": {
              "type": "string"
            },
            "examples": {
              "date": {
                "value": "2023-06-01"
              },
              "datetime": {
                "value": "2023-06-01T05:00:00"
              }
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "end date or datetime. default: default: today 12 pm",
            "required": false,
            "schema": {
              "type": "string"
            },
            "examples": {
              "date": {
                "value": "2023-06-05"
              },
              "datetime": {
                "value": "2023-06-05T20:00:00"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Forecast": {
        "type": "object",
        "properties": {
          "meta": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number",
                "format": "float",
                "example": 55.61
              },
              "lon": {
                "type": "number",
                "format": "float",
                "example": 55.61
              },
              "frequency": {
                "type": "string",
                "enum": [
                  "hourly",
                  "quarterhour"
                ],
                "example": "hourly"
              }
            }
          },
          "values": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "datetime": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2023-06-01T06:00:00.000Z"
                },
                "sumgeneration": {
                  "type": "number",
                  "format": "float",
                  "example": 3504.56
                },
                "price": {
                  "type": "number",
                  "format": "float",
                  "description": "EEX price for one MWh",
                  "example": 80.4
                },
                "temprature": {
                  "type": "number",
                  "format": "float",
                  "description": "temprature in celsius",
                  "example": 15.4
                },
                "windspeed": {
                  "type": "number",
                  "format": "float",
                  "description": "windspeed in m/s",
                  "example": 28.4
                }
              }
            }
          }
        }
      }
    }
  }
}